+++ src/ts/Banners/Managers/BannerCampaignManager.ts
@@ -146,7 +146,6 @@ export class BannerCampaignManager {
     }
 
     private handleError(e: Error, diagnostic: string): Promise<Campaign> {
-        Diagnostics.trigger(diagnostic, e);
         return Promise.reject(e);
     }