+++ tools/release-scripts/Staging/3.create_deployment_prs.sh
@@ -40,4 +40,4 @@ slackjson=$(cat <<EOF
 }
 EOF
 )
-curl -X POST -H 'Content-type: application/json' --data $slackjson https://hooks.slack.com/services/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
+curl -X POST -H 'Content-type: application/json' --data "$slackjson" https://hooks.slack.com/services/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