+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+# Pending
+
+* Install button experiment for rewarded ad videos [#8055](https://github.com/Applifier/unity-ads-webview/pull/8055)
+
+# Thu Feb  7 00:49:07 UTC 2019
+* Fix connect to disconnect in MRAIDAdapterContainer [#8150](https://github.com/Applifier/unity-ads-webview/pull/8150)
+
+# Wed Feb  6 17:45:38 UTC 2019
+
+* Always enable multithreading (Android & iOS) and fix Android multithreading Request API except on group 17 [#8029](https://github.com/Applifier/unity-ads-webview/pull/8029)
+
+# Wed Feb  6 01:02:05 UTC 2019
+
+* Refactor Swipe Class into Separate Classes [#8028](https://github.com/Applifier/unity-ads-webview/pull/8028)
+* Remove banner-auction-request_failed Diagnostic [#8105](https://github.com/Applifier/unity-ads-webview/pull/8105)
+* Decode Url protocols in VastParserStrict [#8058](https://github.com/Applifier/unity-ads-webview/pull/8058)
+
+# Mon Feb  4 23:58:18 UTC 2019
+
+* Remove abTest for performance MRAID CTA handling [#8082](https://github.com/Applifier/unity-ads-webview/pull/8082)
+* Remove Interstitial Layout AB Test [#8083](https://github.com/Applifier/unity-ads-webview/pull/8083)
+* Ignore areControlsVisible in showCallButton [#8060](https://github.com/Applifier/unity-ads-webview/pull/8060)
+* Prefix Release Scripts with Numbers [#8104](https://github.com/Applifier/unity-ads-webview/pull/8104)
+
 # Mon Feb  4 15:46:43 UTC 2019
 
 * Added diagnostics to consent show & send. For investigating a possible live issue. [#8061](https://github.com/Applifier/unity-ads-webview/pull/8061)